Lab 6 part 1

+++ app/controllers/tasks_controller.rb
@@ -13,35 +13,92 @@ class TasksController < ApplicationController
   end
 
   post '/tasks' do
-    task = Task.new(description: params[:description])
+    task = Task.new(description: params[:description], user: current_user)
 
     if task.save
-      redirect "/"
+      if is_json_request?
+        status 201
+        json task
+      else
+        redirect "/"
+      end
     else
-      flash.now[:errors] = task.errors.full_messages.join("; ")
-      erb :"tasks/new.html"
+      if is_json_request?
+        status 400
+        json errors: task.errors.full_messages
+      else
+        flash.now[:errors] = task.errors.full_messages.join("; ")
+        erb :"tasks/new.html"
+      end
     end
   end
 
   get '/tasks/:id' do
-    task = Task.find(params[:id])
-    erb :"tasks/edit.html", locals: { task: task }
-  end
+    task = Task.find_by_id(params[:id])
 
-  put '/tasks/:id' do
-    task = Task.find(params[:id])
-    task.description = params[:description]
-    if task.save
-      redirect "/"
+    if task.nil? or task.user != current_user
+      status 404
+      if is_json_request?
+        halt json({})
+      end
+      halt
+    end
+
+    if is_json_request?
+      json task
     else
-      flash.now[:errors] = task.errors.full_messages.join("; ")
       erb :"tasks/edit.html", locals: { task: task }
     end
   end
 
+  put '/tasks/:id' do
+    task = Task.find_by_id(params[:id])
+
+    if task.nil? or task.user != current_user
+      status 404
+      if is_json_request?
+        halt json({})
+      end
+      halt
+    end
+
+    task.description = params[:description]
+    task.complete = params[:complete]
+
+    if task.save
+      if is_json_request?
+        json task
+      else
+        redirect "/"
+      end
+    else
+      if is_json_request?
+        status 400
+        json errors: task.errors.full_messages
+      else
+        flash.now[:errors] = task.errors.full_messages.join("; ")
+        erb :"tasks/edit.html", locals: { task: task }
+      end
+    end
+  end
+
   delete '/tasks/:id' do
-    task = Task.find(params[:id])
+    task = Task.find_by_id(params[:id])
+
+    if task.nil? or task.user != current_user
+      status 404
+      if is_json_request?
+        halt json({})
+      end
+      halt
+    end
+
     task.destroy!
-    redirect "/"
+
+    if is_json_request?
+      status 204
+    else
+      redirect "/"
+    end
   end
 end
